% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimms-package.R
\docType{data}
\name{bale3g.v1}
\alias{bale3g.v1}
\title{Bale Mountains NDVI3g.v1}
\format{\code{raster::RasterStack}}
\description{
Bale Mountains NDVI3g.v1.
}
\details{
This dataset contains NDVI3g.v1 observations for the Bale Mountains
National Park, southern Ethiopia (Jul 1981 to Dec 2015).
}
