% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createHdr.R
\name{createHdr}
\alias{createHdr}
\title{Create ENVI header file}
\usage{
createHdr(hdr, file)
}
\arguments{
\item{hdr}{Character. If not supplied, defaults to the companion header file
for GIMMS3g binary data. See \code{\link{read.ENVI}} for further information
and section 'Examples' below for required file contents.}

\item{file}{Character. Output filepath, defaults to
\code{paste0(raster::rasterOptions()$tmpdir, "/tmp.hdr")}.}
}
\value{
A filename with the location of the (temporary) header file.
}
\description{
Create an ENVI header file (see \url{http://www.exelisvis.com/docs/ENVIHeaderFiles.html})
to properly import GIMMS binary data into native 'Raster*' format.
}
\examples{
# Create standard GIMMS3G header file
gimms_header <- createHdr()
readLines(gimms_header)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{read.ENVI}}
}

