% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNDVILayers.R
\name{getNDVILayers}
\alias{getNDVILayers}
\title{Get NDVI or flag layers}
\usage{
getNDVILayers(x)
}
\arguments{
\item{x}{\code{list} of 2-layered \code{RasterStack} objects. If not created
from \code{\link{rasterizeGimms}}, make sure that the NDVI (flag) layer is
always the first (second) layer per \code{RasterStack}.}
}
\value{
A \code{RasterStack} object with all NDVI or flag layers in 'x'.
}
\description{
Extract all NDVI or flag layers from the output \code{list} of
\code{\link{rasterizeGimms}}.
}
\author{
Florian Detsch
}

