% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimms-package.R
\docType{data}
\name{kili3g.v0}
\alias{kili3g.v0}
\title{Kilimanjaro NDVI3g.v0}
\format{\code{raster::RasterStack}}
\description{
Kilimanjaro NDVI3g.v0.
}
\details{
This dataset contains NDVI3g.v0 observations for the Kilimanjaro
region, northern Tanzania (Jul 1981 to Dec 2013).
}
