% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldNaming.R
\name{oldNaming}
\alias{oldNaming}
\title{Create Old GIMMS NDVI3g File Names}
\usage{
oldNaming(x, sat = "XX", suffix = "")
}
\arguments{
\item{x}{\code{character}. GIMMS NDVI3g.v1 file names to transform.}

\item{sat}{\code{character}. Satellite number, see 'Details'.}

\item{suffix}{\code{character}. Optional suffix to append to the resulting
file names, e.g. \code{".tif"} for GeoTiff files (see also
\code{\link{writeFormats}}).}
}
\value{
A \code{character} vector of file names following the "traditional" NDVI3g.v0
naming convention.
}
\description{
Transform NDVI3g.v1 file names into "traditional" NDVI3g.v0 naming
convention.
}
\details{
The satellite number was an essential constituent of the NDVI3g.v0 file
naming convention (see 'References'), but has been removed in NDVI3g.v1. In
order to maintain a uniform file naming, the 'sat' argument accepts a single
\code{character} placeholder (defaults to \code{"XX"}) which is inserted in
newer files where no such information is available from NDVI3g.v0.
}
\examples{
fls <- system.file("extdata", "inventory_ecv1.rds", package = "gimms")
fls <- basename(readRDS(fls))

oldNaming(fls[1])
oldNaming(fls[2:3], suffix = ".tif")

}
\references{
\url{https://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
}

