% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\docType{methods}
\name{qualityControl}
\alias{qualityControl}
\alias{qualityControl,RasterStackBrick-method}
\title{Perform quality control on GIMMS NDVI3g data}
\usage{
\S4method{qualityControl}{RasterStackBrick}(x, keep = NULL, filename = "",
  ...)
}
\arguments{
\item{x}{Typically a \code{list} of 2-layered \code{RasterStack} objects
(NDVI and flags) or a single 2-layered \code{RasterStack} object created from
\code{\link{rasterizeGimms}}. Alternatively, a \code{character} vector of raw
GIMMS NDVI3g filenames (see 'Details').}

\item{keep}{\code{integer}. Accepted flag values (see 'Details').}

\item{filename}{\code{character}. Optional output filename(s), see
\code{\link{writeRaster}}. If specified, this must be of the same length as
'x'.}

\item{...}{Arguments passed to \code{\link{writeRaster}}.}
}
\value{
A quality-controlled 'RasterStack'.
}
\description{
Perform quality control on GIMMS NDVI3g data based on the companion flag
information.
}
\details{
If 'keep' is missing, the function will try to determine the GIMMS product
version from which 'x' is derived. In such a case, 'keep' automatically
defaults to \code{1:2} (NDVI3g.v0) or \code{0} (NDVI3g.v1; see 'References').

If 'x' is a \code{character} vector of raw GIMMS NDVI3g filenames,
\code{\link{rasterizeGimms}} is automatically invoked prior to performing
quality control. Note, however, that in such a case, user-specified arguments
'format' and 'overwrite' are not passed to \code{\link{rasterizeGimms}}, but
instead take on default values (\code{format = "GTiff"} and
\code{overwrite = FALSE}). For a finer control, it is hence recommended to
manually run \code{rasterizeGimms(x, ...)} prior to quality control.
}
\references{
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on January 15, 2016).
}
\seealso{
\code{\link{rasterizeGimms}}, \code{\link{writeRaster}}.
}

