% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gimms-package.R
\docType{package}
\name{gimms-package}
\alias{gimms-package}
\alias{gimmspackage}
\title{Download and Process GIMMS3g Data}
\description{
Download and Process GIMMS3g Data
}
\details{
Set of functions to retrieve list of available GIMMS files, download and
extract compressed data, and import binary files into R as standard
'Raster*' objects.
}
\author{
Florian Detsch
}
\references{
Pinzon, JE & Tucker, CJ (2014). A Non-Stationary 1981-2012 AVHRR
NDVI3g Time Series. Remote Sensing, 6(8), 6929-6960. Available online at
\url{http://www.mdpi.com/2072-4292/6/8/6929/htm}.
}
\keyword{package}

