% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimms-package.R
\docType{package}
\name{gimms-package}
\alias{gimms-package}
\alias{gimmspackage}
\title{Download and Process GIMMS3g Data}
\description{
Download and Process GIMMS3g Data
}
\details{
We provide a set of functions to retrieve information about GIMMS NDVI3g
files currently available online; download and re-arrange the bi-monthly
datasets according to creation time; import downloaded files from native
binary (ENVI) format directly into R based on the widely applied 'raster'
package; and calculate monthly value composites (e.g. maximum value
composites, MVC) from the bi-monthly input data.
}
\author{
Florian Detsch
}
\references{
Pinzon, JE & Tucker, CJ (2014). A Non-Stationary 1981-2012 AVHRR
NDVI3g Time Series. Remote Sensing, 6(8), 6929-6960. Available online at
\url{http://www.mdpi.com/2072-4292/6/8/6929/htm}.

Pinzon, JE & Tucker, CJ (2016). A Non-Stationary 1981-2015 AVHRR
NDVI3g.v1 Time Series: an update. In preparation for submission to Remote
Sensing.
}
\keyword{package}

