% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\docType{methods}
\name{qualityControl}
\alias{qualityControl}
\alias{qualityControl,RasterStackBrick-method}
\alias{qualityControl,character-method}
\title{Perform quality control on GIMMS NDVI3g data}
\usage{
\S4method{qualityControl}{RasterStackBrick}(x, flag, keep = 1:2, cores = 1L,
  filename = "", ...)

\S4method{qualityControl}{character}(x, flag, keep = 1:2, cores = 1L,
  filename = "", ...)
}
\arguments{
\item{x}{'RasterStack' with NDVI values or 'character' vector of ENVI binary
files. If the latter applies, \code{flag} is created internally via
\code{\link{rasterizeGimms}}, but note that this comes at a high memory cost
since the rasterized images cannot be written to disk at the moment.}

\item{flag}{'RasterStack' with flag values. Ignored if \code{x} is a
'character' vector.}

\item{keep}{Integer. Accepted flag values, defaults to \code{1:2}
(\emph{i.e.}, 'good' values; see 'References'.)}

\item{cores}{Integer. Number of cores for parallel computing.}

\item{filename}{Character. Optional output filename(s); see
\code{\link{writeRaster}}. If \code{cores > 1}, the number of supplied
filenames must match up with the number of layers in \code{x}.}

\item{...}{Further arguments passed on to \code{\link{writeRaster}} (except
for 'bylayer' and 'suffix').}
}
\value{
A quality-controlled 'RasterStack'.
}
\description{
Perform quality control on GIMMS NDVI3g data based on the companion flag
information.
}
\examples{
\dontrun{
## Download sample data
gimms_dir <- paste0(getwd(), "/data")

gimms_files <- downloadGimms(x = as.Date("2000-01-01"),
                             y = as.Date("2000-06-30"), dsn = gimms_dir)

## Rasterize files
gimms_qc <- qualityControl(gimms_files)
plot(gimms_qc[[1:4]])
}

}
\author{
Florian Detsch
}
\references{
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on January 15, 2016).
}
\seealso{
\code{\link{rasterizeGimms}}, \code{\link{writeRaster}}.
}

