% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\docType{methods}
\name{qualityControl}
\alias{qualityControl}
\alias{qualityControl,RasterStackBrick-method}
\title{Perform quality control on GIMMS NDVI3g data}
\usage{
\S4method{qualityControl}{RasterStackBrick}(x, keep = NULL, filename = "",
  ...)
}
\arguments{
\item{x}{A single 2-layered \code{RasterStack} object (NDVI and flags).}

\item{keep}{\code{integer}. Accepted flag values (see 'Details').}

\item{filename}{\code{character}. Optional output filename(s), see
\code{\link{writeRaster}}. If specified, this must be of the same length as
'x'.}

\item{...}{Arguments passed to \code{\link{writeRaster}}.}
}
\value{
A quality-controlled 'RasterLayer' object.
}
\description{
Perform quality control on GIMMS NDVI3g data based on the companion flag
information.
}
\details{
If 'keep' is missing, the function will automatically skip quality control
and return the input object. See the 'References' section for more
information about product-specific flag values.
}
\examples{
## see 'Examples' section in ?rasterizeGimms

}
\references{
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on January 15, 2016).
}
\seealso{
\code{\link{overlay}}.
}

