% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\docType{methods}
\name{qualityControl}
\alias{qualityControl}
\alias{qualityControl,RasterStackBrick-method}
\alias{qualityControl,list-method}
\title{Perform quality control on GIMMS NDVI3g data}
\usage{
\S4method{qualityControl}{RasterStackBrick}(x, keep = 0:2, ...)

\S4method{qualityControl}{list}(x, keep = 0:2, cores = 1L, ...)
}
\arguments{
\item{x}{Typically a \code{list} of 2-layered \code{RasterStack} objects
(NDVI and flags) or a single 2-layered \code{RasterStack} object created from
\code{\link{rasterizeGimms}}. If a \code{character} vector of filenames is
supplied, it is assumed that these represent NDVI layers, and hence, flag
layers must be supplied to 'y'.}

\item{keep}{Integer. Accepted flag values, defaults to \code{1:2}
(\emph{i.e.}, 'good' values; see 'References'.)}

\item{...}{Arguments passed on to \code{\link{writeRaster}}.}

\item{cores}{Integer. Number of cores for parallel computing.}

\item{y}{\code{character} vector of filenames, only considered if 'x' is a
\code{character} object as well. See 'Details'}
}
\value{
A quality-controlled 'RasterStack'.
}
\description{
Perform quality control on GIMMS NDVI3g data based on the companion flag
information.
}
\details{
(to be continued...)
}
\examples{
\dontrun{
## Download sample data
gimms_dir <- paste0(getwd(), "/data")

gimms_files <- downloadGimms(x = as.Date("2000-01-01"),
                             y = as.Date("2000-06-30"), dsn = gimms_dir)

## Rasterize files
gimms_qc <- qualityControl(gimms_files)
plot(gimms_qc[[1:4]])
}

}
\author{
Florian Detsch
}
\references{
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on January 15, 2016).
}
\seealso{
\code{\link{rasterizeGimms}}, \code{\link{writeRaster}}.
}

