% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rasterizeGimms.R
\name{rasterizeGimms}
\alias{rasterizeGimms}
\title{Rasterize GIMMS3g binary data}
\usage{
rasterizeGimms(x, hdr, water2na = TRUE, nodata2na = TRUE, scaling = TRUE,
  filename = "", ...)
}
\arguments{
\item{x}{Character. Vector of local filepaths.}

\item{hdr}{Character. Companion header file passed on to
\code{\link{caTools::read.ENVI}}. If missing, the standard header file for
GIMMS3g binary data as created by \code{\link{createHdr}} will be used.}

\item{water2na}{Logical. If \code{TRUE} (default), pixels with 'mask-water'
value (-10000) will be discarded. See also
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}.}

\item{nodata2na}{Logical. If \code{TRUE} (default), pixels with 'mask-nodata'
value (-5000) will be discarded.}

\item{scaling}{Logical. If \code{TRUE} (default), initial values will be
scaled by a factor of 1/10000.}

\item{filename}{Character. Optional output filename, see
\code{\link{raster::writeRaster}}.}

\item{...}{Further arguments passed on to \code{\link{raster::writeRaster}}.}
}
\value{
If 'x' is a single filename, an object of class 'RasterLayer'; <br>
if 'x' is a vector of filenames, an object of class 'RasterStack'.
}
\description{
Import GIMMS3g binary data into R as 'Raster*' objects based on a companion
header file.
}
\examples{
\dontrun{
## Download GIMMS3g binary data from 2000-2005 (this might take some time...)
gimms_files <- downloadGimms(x = 2000, y = 2005,
                             dsn = paste0(getwd(), "/data"))

## Rasterize downloaded GIMMS files from 2000 (this might take even longer...)
gimms_raster <- rasterizeGimms(x = gimms_files[1:24])

plot(rst_gimms)
}
}
\author{
Florian Detsch
}
\seealso{
\code{\link{createHdr}}, \code{\link{caTools::read.ENVI}},
\code{\link{raster::raster}}, \code{\link{raster::writeRaster}}.
}

