% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeGimms.R
\name{rasterizeGimms}
\alias{rasterizeGimms}
\title{Rasterize GIMMS NDVI3g data}
\usage{
rasterizeGimms(x, ext = NULL, snap = "out", keep = NULL, split = FALSE,
  cores = 1L, filename = "", ...)
}
\arguments{
\item{x}{\code{character}. Vector of local filepaths.}

\item{ext}{\code{Extent}, or any object from which an \code{Extent} can be
extracted, see \code{\link[raster]{crop}}.}

\item{snap}{\code{character}, defaults to "out". Other available options are
"in" and "near", see \code{\link[raster]{crop}}.}

\item{keep}{\code{integer}. Flag values of NDVI3g pixels to spare during
quality control. Pixels with non-included flag values are set to \code{NA}.
If not specified (i.e., \code{NULL}; default), quality control is skipped.}

\item{split}{\code{logical}, defaults to \code{FALSE}. If \code{TRUE}, a
\code{list} of \code{RasterStack} objects (of \code{length(x)}) is returned
rather than a single \code{RasterStack}.}

\item{cores}{\code{integer}. Number of cores for parallel computing.}

\item{filename}{\code{character}.}

\item{...}{Arguments passed to \code{\link{writeRaster}}.}
}
\value{
If \code{split = TRUE}, a list of NDVI3g \code{RasterStack} objects
corresponding to the files specified in 'x'; else a single NDVI3g
\code{RasterStack} object.
}
\description{
Import GIMMS NDVI3g (binary or NetCDF) data into R as \code{Raster*} objects.
}
\examples{
\dontrun{
## Download NDVI3g.v1 sample data
gimms_files <- downloadGimms(x = as.Date("2000-01-01"),
                             y = as.Date("2000-12-31"))

## Extent for clipping
shp <- getData("GADM", country = "DEU", level = 0, path = tmpDir())

## Rasterize without quality control
gimms_raster <- rasterizeGimms(x = gimms_files,
                               ext = shp) # clipping
plot(gimms_raster[[1]])
lines(shp)

## Rasterize with quality control
gimms_rasterq <- rasterizeGimms(x = gimms_files,
                                ext = shp, # clipping
                                keep = 0)  # quality control
plot(gimms_rasterq[[1]])
lines(shp)
}

}
\author{
Florian Detsch
}
\references{
\url{https://nex.nasa.gov/nex/projects/1349/wiki/general_data_description_and_access/}
(accessed on 11/04/2016).
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on 11/04/2016).
}
\seealso{
\code{\link[raster]{crop}}, \code{\link{writeRaster}}.
}

