% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeGimms.R
\name{rasterizeGimms}
\alias{rasterizeGimms}
\title{Rasterize GIMMS NDVI3g data}
\usage{
rasterizeGimms(x, dsn = dirname(x), water2na = TRUE, nodata2na = TRUE,
  scaling = TRUE, cores = 1L, format = "GTiff", overwrite = FALSE)
}
\arguments{
\item{x}{\code{character}. Vector of local filepaths.}

\item{dsn}{\code{character}. Destination folder for rasterized files.
Defaults to \code{dirname(x)} if not further specified.}

\item{water2na}{\code{logical}. Determines whether or not to discard pixels
with 'mask-water' value (see 'References').}

\item{nodata2na}{\code{logical}. Determines whether or not to discard pixels
with 'mask-nodata' value (see 'References').}

\item{scaling}{\code{logical}. If \code{TRUE} (default), scaling is enabled
and both the NDVI and flag layer per file in 'x' are returned.}

\item{cores}{\code{integer}. Number of cores for parallel computing.}

\item{format}{\code{character}. Output file type, see
\code{\link{writeRaster}}.}

\item{overwrite}{\code{logical}. If \code{TRUE}, existing output files will
be overwritten, see \code{\link{writeRaster}}.}
}
\value{
If \code{scaling = TRUE} and \code{length(x) > 1}, a \code{list} of
\code{length(x)} consisting of 2-layered \code{RasterStack} objects (NDVI and
flags);

Else if \code{scaling = TRUE} and \code{length(x) == 1}, a 2-layered
\code{RasterStack} objects (NDVI and flags);

Else if \code{scaling = FALSE}, a \code{list} of \code{length(x)} consisting
of \code{RasterLayer} objects with raw GIMMS NDVI3g values.
}
\description{
Import GIMMS NDVI3g (binary or NetCDF) data into R as \code{Raster*} objects.
}
\examples{
\dontrun{
## Download sample data
gimms_files <- downloadGimms(x = as.Date("2000-01-01"),
                             y = as.Date("2000-06-30"))

## Rasterize files
gimms_raster <- rasterizeGimms(x = gimms_files)
gimms_raster
}

}
\author{
Florian Detsch
}
\references{
\url{https://nex.nasa.gov/nex/projects/1349/wiki/general_data_description_and_access/}
(accessed on 11/04/2016).
\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/00READMEgeo.txt}
(accessed on 11/04/2016).
}
\seealso{
\code{\link{writeRaster}}.
}

