% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantTau.R
\name{significantTau}
\alias{significantTau}
\title{Compute (pre-whitened) Kendall's tau}
\usage{
significantTau(x, p = 0.001, prewhitening = FALSE, df = FALSE, ...)
}
\arguments{
\item{x}{A 'numeric' vector.}

\item{p}{Significance level to be tested.}

\item{prewhitening}{'logical'. If \code{TRUE}, pre-whitening is applied prior
to the Mann-Kendall trend test.}

\item{df}{'logical'. If \code{TRUE}, a 'data.frame' holding the value of
Kendall's tau and the referring significance level.}

\item{...}{Further arguments passed on to \code{\link{zyp.trend.vector}}.}
}
\value{
If \code{df = FALSE} (default) and \code{p} was not exceeded, a
single 'numeric'; else a 'data.frame' with Kendall's tau and the referring
significance level.
}
\description{
Apply the Mann-Kendall trend test (Mann, 1945) to a series of observations
and return Kendall's tau (Kendall, 1938) based on a predefined significance
level. In contrast to other readily available implementations, it is left to
the user to decide whether or not to apply pre-whitening as described in the
\strong{zyp} package vignette (Bronaugh and Werner, 2013).
}
\examples{
## Example taken from ?Kendall::MannKendall
library(Kendall)
data(PrecipGL)
plot(PrecipGL)

## Mann-Kendall trend test without pre-whitening
significantTau(PrecipGL, p = 0.001)

## Mann-Kendall trend test with pre-whitening
significantTau(PrecipGL, p = 0.001, prewhitening = TRUE)

}
\author{
Florian Detsch
}
\references{
Kendall, M.G. (1938). A new measure of rank correlation. \emph{Biometrika}
30(1/2), 81-93, doi: 10.2307/2332226. Available online at
\url{http://www.jstor.org/stable/2332226} (accessed 2015-11-06).

Mann, H.B. (1945). Nonparametric tests against trend. \emph{Econometrica}
13(3), 245-259, doi: 10.2307/1907187. Available online at
\url{http://www.jstor.org/stable/1907187} (accessed 2015-11-06).

Zhang, X., Vincent, L.A., Hogg, W.D. and A. Niitsoo (2000). Temperature and
Precipitation Trends in Canada during the 20th Century. \emph{Atmosphere-Ocean}
38(3), 395-429, doi: 10.1080/07055900.2000.9649654. Available online at
\url{http://www.tandfonline.com/doi/abs/10.1080/07055900.2000.9649654}
(accessed 2015-11-06).

Yue, S., Pilon, P., Phinney, B. and G. Cavadias (2002). The influence of
autocorrelation on the ability to detect trend in hydrological series.
\emph{Hydrological Processes} 16, 1807-1829, doi: 10.1002/hyp.1095.
Available online at \url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.1095/abstract}
(accessed 2015-11-06).
}
\seealso{
\code{\link{MannKendall}}, \code{\link{zyp.trend.vector}}.
}

