% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInventory.R
\name{updateInventory}
\alias{updateInventory}
\title{Update GIMMS NDVI3g file inventory}
\usage{
updateInventory(server = c("ecocast", "nasanex"), version = 1L)
}
\arguments{
\item{server}{\code{character}. Specifies the remote server to use. Currently
available options are \code{"ecocast"} (default) and \code{"nasanex"}.}

\item{version}{\code{integer} (or any other convertible class), defaults to
\code{1L}. Specifies desired GIMMS NDVI3g product version, see 'Details' and
'References'. Currently ignored if \code{server = "nasanex"}.}
}
\value{
A vector of online filepaths.
}
\description{
Download the latest version of the GIMMS NDVI3g file inventory from the NASA
Ames Ecological Forecasting Lab (ECOCAST) or NASA Earth Exchange (NEX) Amazon
AWS. If none of the specified endpoints is reachable (e.g., if there is no
active internet connection), the latest local version of the file inventory
is used.
}
\details{
Only recently, GIMMS NDVI3g.v1 has been released which is currently available
from ECOCAST until end 2015 and comes in NetCDF (\code{.nc4}) format. In
contrast, NDVI3g.v0 comes in ENVI raw binary format is currently available
from ECOCAST (NASANEX) until end 2013 (2012).
}
\examples{
updateInventory()

}
\author{
Florian Detsch
}
\references{
ECOCAST (2016). Available online (accessed on 27 October 2016):
\itemize{
\item{NDVI3g.v1 (\code{.nc4}, until end 2015)}
{\url{https://ecocast.arc.nasa.gov/data/pub/gimms/3g.v1/}}
\item{NDVI3g.v0 (raw binary, until end 2013)}
{\url{https://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/}}
}

NASANEX (2016). Available online (accessed on 27 October 2016):
\itemize{
\item{NDVI3g.v0 (raw binary, until end 2012)}
\url{https://nasanex.s3.amazonaws.com/}
}
}
\seealso{
\code{\link{rearrangeFiles}}.
}

