% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInventory.R
\name{updateInventory}
\alias{updateInventory}
\title{Update GIMMS 3G file inventory}
\usage{
updateInventory(sort = TRUE, quiet = TRUE)
}
\arguments{
\item{sort}{'logical'. If \code{TRUE}, the list of available files is sorted
by date prior to return.}

\item{quiet}{'logical'. If \code{FALSE}, some details about the online file
retrieval are printed to the console.}
}
\value{
A 'character' vector of online filepaths.
}
\description{
Download the latest version of the GIMMS 3G file inventory from NASA FTP
server (\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/}) or, if
server is not accessible (e.g. due to missing internet connection), load
local version of file inventory.
}
\examples{
updateInventory()

}
\author{
Florian Detsch
}

