% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInventory.R
\name{updateInventory}
\alias{updateInventory}
\title{Update GIMMS 3G file inventory}
\usage{
updateInventory(sort = FALSE)
}
\arguments{
\item{sort}{Logical. If \code{TRUE} (default), the available files are sorted
by date prior to return.}
}
\value{
A vector of online filepaths.
}
\description{
Download the latest version of the GIMMS 3G file inventory from NASA FTP
server (\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/}) or, if
server is not accessible (e.g. due to missing internet connection), load
local version of file inventory.
}
\examples{
updateInventory()

}
\author{
Florian Detsch
}

